engine/layoutgen/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
Bottleneck
Brotli
certifi==2025.4.26
charset-normalizer==3.4.2
colorama
comm
contourpy
cycler
debugpy
decorator
exceptiongroup
executing
Flask==2.3.3
Flask-CORS==4.0.0
fonttools
frozenlist==1.6.2
idna==3.10
ipykernel
ipython
jedi
jupyter_client
jupyter_core
kiwisolver
matplotlib==3.10.0
matplotlib-inline
mkl-service==2.4.0
mkl_fft
mkl_random
multidict==6.4.4
nest-asyncio
networkx==3.1
numexpr
numpy==1.24.3
packaging
pandas
parso
pillow
platformdirs
prompt-toolkit
propcache==0.3.1
psutil
pure-eval
Pygments
pyparsing
PyQt6==6.7.1
PyQt6_sip
python-dateutil
python-dotenv==1.1.0
pytz
pywin32==308
pyzmq
requests==2.32.3
sip
six
stack-data
tomli
tornado
traitlets
typing_extensions
tzdata
unicodedata2
urllib3==2.4.0
wcwidth
yarl==1.20.0